AME-7894 AME-7896 CREST Debug Endpoint and SSOAdmin command
Rest endpoint and a SSOAdmin command for recording OpenAM. It will generate a record bundle that will contain:

- debug logs during the time-frame of the record, in the requested debug level.
- thread dump generated during the recording
- a config export, generated at the beginning of the recording
- an info.json, with the environment info, JVM info and more
- a RecordHistory.txt, with the record time history

This record bundle will help investigating in OpenAM issue. The communication between the client and the support will be improved, client will send this bundle on each ticket.
It's also a good starting point for a support bundle format, which will allow us to start building analyze tools based on it.