AME-8337 CR-8191 Convert themeConfig.json to ThemeConfiguration.js
By using a JavaScript module, we give the user more flexibilty in
defining their theming config. It also allows us to include comments
to act as documentation and to provide a simpler syntax for
regular expressions (as these can be written inline in JavaScript).

This change also includes significant refactoring of the
ThemeManager to simplify it.