AME-8327 Move the majority of build tasks from Maven to Grunt
A JavaScript build process is more manageable and is compatible
with far more JavaScript tooling than the existing Maven process.
As an added benefit, Node.js executes the build an order of magnitude
faster than the Java-based Rhino runtime.

In order to provide a version of Node.js for the JavaScript build
to use, this change introduces the use of frontend-maven-plugin, which
downloads and executes Node.js.

Maven is currently still responsible for the unpacking of the
project dependencies and the packaging of the final artefact.

The only change in behaviour in this change is the replacement of
JSLint with ESLint. This is because the version of JSLint used by
the Maven build was 2011-03-07, which is not supported by the
grunt wrapper.