OPENAM-6734: Prevent shutdown race condition between embedded OpenDJ and OpenAM persistent search restart
See OPENAM-6734 description for context on the original issue and a useful stacktrace.

This is not a full fix for all cases where getTimerPool() might be called during shutdown without proper error handling for the runtime exception. (There are 20+ candidates, but to be fair a lot of them seem far less likely to ever crop up as an issue in comparison to this one).

I've raised OPENAM-7076 to track the overall issue.