Replace QUnit with Karma/Mocha/Chai
Karma provides a web server for hosting the tests, which removes
the requirement for an application server to deploy to. This makes
it possible to easily integrate the unit tests into the build. It
also watches the source files and automatically reruns the tests.

Mocha and Chai provide a modern testing experience, with a BDD
testing style and comprehensive assertions.

As part of this change, the QUnit tests that were functional-style
tests have been removed. The plan is to eventually replace them
with true functional Selenium tests.