Remove global comments
ESLint allows us to specify what it calls "environments". This
controls which globals it allows files to reference. By enabling
the "amd" environment, we can happily reference define and
require, which are made available through our use of RequireJS.

Using this mechanism is less error prone and verbose than the
continued use of global comments.