OPENAM-7058 Avoid converting session expiry times
The conversion from session expiry time to time left and back again was
introducing error. The time left is relative to the time at which it was
computed, and then later converted back into an absolute time using a
different relative point, leading to a different time than the original
expiry time. The difference was non-deterministic and depended on the
speed of execution of the code.

When stateless sessions are enabled, this error in the expiry time
caused blacklisting to fail.