Allow running with a volume for configuration
`$OPENAM_DATA_DIR` does not exist in the image. Because of this, any
volume mounted to that location will be owned by `root` and unwritable
by `$OPENAM_USER`.

By creating this directory (owned by `$OPENAM_USER`), a newly created
volume mounted on `$OPENAM_DATA_DIR` will get its permissions set to
whatever the underlying directory is set to, so OpenAM can write to it.